+++ examples/tokio_fetch_stream.rs
@@ -35,7 +35,7 @@ use io_imap::{
     session::*,
     types::mailbox::Mailbox,
 };
-use pimalaya_stream::sasl::{Sasl, SaslPlain};
+use io_sasl::{mechanism::Sasl, rfc4616::plain::SaslPlainCreds};
 use rustls::{ClientConfig, pki_types::ServerName};
 use rustls_platform_verifier::ConfigVerifierExt;
 use tokio::{
@@ -73,7 +73,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
         .install_default()
         .ok();
 
-    let sasl = SaslPlain {
+    let sasl = SaslPlainCreds {
         authzid: None,
         authcid: env::var("LOGIN")?,
         passwd: env::var("PASSWORD")?.into(),